code_share/requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.2
django-test-plus==1.0.18
djangorestframework==3.6.3
flake8==3.4.1
gunicorn==19.7.1
mccabe==0.6.1
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
pytest==3.2.0
pytest-django==3.1.2
<<<<<<< HEAD
python-decouple==3.0
=======
>>>>>>> 241842cb9400a3562e848e9d76455a730741dead
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
